Created use cases and structure to rest

+++ internal/domain/entities/metric.go
@@ -3,5 +3,5 @@ package entities
 type Metric struct {
 	Name      string
 	Attribute string
-	Value     int
+	Value     float64
 }
