+++ internal/usecases/contracts/istandard_deviation_use_case.go
@@ -5,5 +5,5 @@ import (
 )
 
 type StandardDeviationUseCase interface {
-	Execute(ctx context.Context, variance int) int
+	Execute(ctx context.Context, variance int) float64
 }
